pyproject.toml
[project]
name = "Photolog"
version = "0.1.0"
description = "A simple photolog web application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.3",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.0.0",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.9.0",
    "python-jose>=3.3.0",
    "sqlmodel>=0.0.22",
    "typer>=0.12.5",
]
